conf-tweaks: Rename from pmos-tweaks
The name is being changed for two reasons:

- While this scheme originated in postmarketOS, it is not strictly
  related or dependent on postmarketOS.
- The name postmarketOS is due to be changed soon, but actually
  changing it will take a while. Waiting for a new name before landing
  this doesn't make much sense.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh-mobile-settings/-/merge_requests/242>
