+++ src/mobile-settings-window.c
@@ -17,9 +17,9 @@
 #include "ms-feedback-panel.h"
 #include "ms-plugin-panel.h"
 
-#include "pmos-tweaks/ms-tweaks-page-builder.h"
-#include "pmos-tweaks/ms-tweaks-parser.h"
-#include "pmos-tweaks/ms-tweaks-utils.h"
+#include "conf-tweaks/ms-tweaks-page-builder.h"
+#include "conf-tweaks/ms-tweaks-parser.h"
+#include "conf-tweaks/ms-tweaks-utils.h"
 
 #include <glib/gi18n.h>
