+++ tests/test-tweaks-gtk-utils.c
@@ -6,7 +6,7 @@
  * Author: Stefan Hansson <newbyte@postmarketos.org>
  */
 
-#include "pmos-tweaks/ms-tweaks-gtk-utils.h"
+#include "conf-tweaks/ms-tweaks-gtk-utils.h"
 
 
 static void
